% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{data}
\alias{data}
\title{A dataset used for illustration.}
\description{
This dataset is used to compare the performance of C function \code{vaccR}) and C++ function \code{vaccC}.
}
\examples{
\dontrun{
data(data)
attach(data)
tm <- microbenchmark::microbenchmark(
  vR = vaccR(age,female,ily),
  vC = vaccC(age,female,ily)
)
print(summary(tm)[,c(1,3,5,6)])
}
}
