% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{monte_carlo_pi}
\alias{monte_carlo_pi}
\title{Monte Carlo Estimation of Pi}
\usage{
monte_carlo_pi(n)
}
\arguments{
\item{n}{An integer specifying the number of random points to generate}
}
\value{
An estimate of Pi
}
\description{
Use Monte Carlo simulation to estimate the value of Pi
}
