% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{optimal_matchit_matching}
\alias{optimal_matchit_matching}
\title{Optimal Matching using MatchIt}
\usage{
optimal_matchit_matching(data, treatment, covariates)
}
\arguments{
\item{data}{A data frame containing the treatment variable and covariates.}

\item{treatment}{A character string specifying the treatment variable (binary: 0 or 1).}

\item{covariates}{A character vector of covariate names used for propensity score calculation.}
}
\value{
A data frame with matched data from `matchit` using optimal matching.
}
\description{
This function uses `matchit` package to perform optimal matching based on propensity scores.
}
