% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qr_decomposition}
\alias{qr_decomposition}
\title{QR Decomposition}
\usage{
qr_decomposition(A)
}
\arguments{
\item{A}{A numeric matrix}
}
\value{
A list containing matrices Q and R
}
\description{
Compute the QR decomposition of a matrix A
}
